fixed overflow

+++ source/renderer_helpers.c
@@ -13,11 +13,11 @@ draw_pixel(vec2_t position , u32 color)
 
 // Draws a rectangle on the screen at a certain coordinate X and Y with a Color.
 internal void
-draw_rect(u16 x, u16 y, u16 w, u16 h, u32 color)
+draw_rect(s32 x, s32 y, s32 w, s32 h, u32 color)
 {
-    for(u16 height_index = y; height_index < y + h; ++height_index)
+    for(s32 height_index = y; height_index < y + h; ++height_index)
     {
-        for(u16 width_index = x; width_index < x + w; ++width_index)
+        for(s32 width_index = x; width_index < x + w; ++width_index)
         {
             if(width_index < window_width && height_index < window_height)
             {
